feat: no-flickering theme
